Removing "B Council" references
bcouncil.html is removed.

Not shown in this commit is the removal of `bcouncil/*`. That folder
and the associated HTML file have been zipped as bcouncil.zip,
which will be provided to the webmaster with the full source code.

Based on analysis of the "B Council" documents and meeting notes,
it ran from 2003-2005 and was a shadow council within CMRF. Talk to
old Rennies about what it was, and if you find out, throw a commit in
with more information, please.

- Ben Keith.146